+++ src/main/java/web/WebUIApplication.java
@@ -11,8 +11,8 @@ import chess.domain.game.state.BlackWin;
 import chess.domain.game.state.State;
 import chess.domain.game.state.WhiteWin;
 import chess.domain.piece.Piece;
-import database.ChessGameDao;
-import database.RoomDao;
+import web.dao.ChessGameDao;
+import web.dao.RoomDao;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import spark.ModelAndView;
